+++ src/IDBKeyRange.js
@@ -12,7 +12,7 @@ import * as util from './util.js';
  */
 function IDBKeyRange (lower, upper, lowerOpen, upperOpen) {
     if (lower === undefined && upper === undefined) {
-        throw new TypeError('Both arguments to the key range method cannot be undefined');
+        throw createDOMException('DataError', 'Both arguments to the key range method cannot be undefined');
     }
     if (lower !== undefined) {
         Key.convertValueToKey(lower);
@@ -32,21 +32,36 @@ function IDBKeyRange (lower, upper, lowerOpen, upperOpen) {
     this.__upperOpen = !!upperOpen;
 }
 IDBKeyRange.prototype.includes = function (key) {
+    if (!arguments.length) {
+        throw new TypeError('IDBKeyRange.includes requires a key argument');
+    }
     Key.convertValueToKey(key);
     return Key.isKeyInRange(key, this);
 };
 
 IDBKeyRange.only = function (value) {
+    if (!arguments.length) {
+        throw new TypeError('IDBKeyRange.only requires a value argument');
+    }
     return new IDBKeyRange(value, value, false, false);
 };
 
 IDBKeyRange.lowerBound = function (value, open) {
+    if (!arguments.length) {
+        throw new TypeError('IDBKeyRange.lowerBound requires a value argument');
+    }
     return new IDBKeyRange(value, undefined, open, true);
 };
 IDBKeyRange.upperBound = function (value, open) {
+    if (!arguments.length) {
+        throw new TypeError('IDBKeyRange.upperBound requires a value argument');
+    }
     return new IDBKeyRange(undefined, value, true, open);
 };
 IDBKeyRange.bound = function (lower, upper, lowerOpen, upperOpen) {
+    if (arguments.length <= 1) {
+        throw new TypeError('IDBKeyRange.bound requires lower and upper arguments');
+    }
     return new IDBKeyRange(lower, upper, lowerOpen, upperOpen);
 };
 IDBKeyRange.prototype[Symbol.toStringTag] = 'IDBKeyRange';